[PORT-21] Move the sidebarItems array to a different file
