+++ src/components/template/sidebar.vue
@@ -1,32 +1,5 @@
 <script setup lang="ts">
-import { ref } from "vue";
-
-const sidebarItems = [
-    {
-        id: 1,
-        title: "HOME",
-        linkTo: "#home",
-        icon: "mdi:home",
-    },
-    {
-        id: 2,
-        title: "WORK",
-        linkTo: "#work",
-        icon: "mdi:work",
-    },
-    {
-        id: 3,
-        title: "ABOUT",
-        linkTo: "#about",
-        icon: "mdi:info",
-    },
-    {
-        id: 4,
-        title: "CONTACT",
-        linkTo: "#contact",
-        icon: "mdi:mail",
-    },
-];
+import sidebarItems from '@/data/sidebar-items';
 
 const githubLink = "https://github.com/jasonchuacodes";
 const linkedInlink = "https://www.linkedin.com/in/jason-clyde-chua-65a322243/";
