+++ src/data/sidebar-items.js
@@ -1,0 +1,28 @@
+const sidebarItems = [
+  {
+      id: 1,
+      title: "HOME",
+      linkTo: "#home",
+      icon: "mdi:home",
+  },
+  {
+      id: 2,
+      title: "WORK",
+      linkTo: "#work",
+      icon: "mdi:work",
+  },
+  {
+      id: 3,
+      title: "ABOUT",
+      linkTo: "#about",
+      icon: "mdi:info",
+  },
+  {
+      id: 4,
+      title: "CONTACT",
+      linkTo: "#contact",
+      icon: "mdi:mail",
+  },
+];
+
+export default sidebarItems;
